Create OpeningHoursContent
