+++ src/pages/steps/CalendarStep/FixedDays.tsx
@@ -1,4 +1,3 @@
-import { format } from 'date-fns';
 import { ChangeEvent, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -6,17 +5,14 @@ import { useHolidaysWithToggle } from '@/hooks/api/holidays';
 import { FeatureFlags, useFeatureFlag } from '@/hooks/useFeatureFlag';
 import { OpeningHours } from '@/types/Offer';
 import { Box } from '@/ui/Box';
-import { Button, ButtonVariants } from '@/ui/Button';
+import { ButtonVariants } from '@/ui/Button';
 import { DatePeriodPicker } from '@/ui/DatePeriodPicker';
-import { Inline } from '@/ui/Inline';
-import { List } from '@/ui/List';
 import { Modal, ModalSizes, ModalVariants } from '@/ui/Modal';
 import { RadioButtonWithLabel } from '@/ui/RadioButtonWithLabel';
 import { Stack } from '@/ui/Stack';
 import { Text } from '@/ui/Text';
 import { colors } from '@/ui/theme';
 
-import type { SupportedLanguage } from '../../../i18n';
 import {
   CalendarState,
   useCalendarSelector,
@@ -28,6 +24,7 @@ import { CalendarOpeninghoursModal } from './CalendarOpeninghoursModal';
 import { CalendarOpeninghoursModalLegacy } from './CalendarOpeninghoursModalLegacy';
 import type { ClosingPeriodData } from './ClosingPeriod';
 import type { DeviatingPeriodData } from './DeviatingPeriod';
+import { OpeningHoursContent } from './OpeningHoursContent';
 
 const FixedDayOptions = {
   PERMANENT: 'permanent',
@@ -58,8 +55,7 @@ export const FixedDays = ({
   onChangeClosingPeriods,
   initialClosingPeriods,
 }: FixedDaysProps) => {
-  const { t, i18n } = useTranslation();
-  const lang = i18n.language as SupportedLanguage;
+  const { t } = useTranslation();
   const [isBoaEnabled] = useFeatureFlag(FeatureFlags.BOA);
 
   const [
@@ -79,16 +75,6 @@ export const FixedDays = ({
   const endDate = useCalendarSelector((state) => state.context.endDate);
 
   const { apiHolidays, onShowHolidaysChange } = useHolidaysWithToggle();
-
-  const openingHours = useCalendarSelector(
-    (state) => state.context.openingHours,
-  );
-
-  const hasOpeningHours = openingHours.length > 0;
-  const hasBoaContent =
-    hasOpeningHours ||
-    (initialAdjustedDays?.length ?? 0) > 0 ||
-    (initialClosingPeriods?.length ?? 0) > 0;
 
   const handleChangeOption = (event: ChangeEvent<HTMLInputElement>) => {
     const value = event.target.value;
@@ -113,224 +99,6 @@ export const FixedDays = ({
     handleChangeOpeningHours([]);
     setIsDeleteConfirmVisible(false);
   };
-
-  const openingHoursContent =
-    isBoaEnabled && hasBoaContent ? (
-      <Stack spacing={3}>
-        <Inline justifyContent="space-between" alignItems="center">
-          <Text fontWeight="bold">
-            {t(
-              isBoaEnabled
-                ? 'create.calendar.fixed_days.hours'
-                : 'create.calendar.fixed_days.opening_hours',
-            )}
-          </Text>
-          <Inline spacing={3}>
-            <Button
-              variant={ButtonVariants.LINK}
-              onClick={() => setIsCalendarOpeninghoursModalVisible(true)}
-            >
-              {t('create.calendar.fixed_days.overview.edit')}
-            </Button>
-            <Button
-              variant={ButtonVariants.LINK_DANGER}
-              onClick={() => setIsDeleteConfirmVisible(true)}
-            >
-              {t('create.calendar.fixed_days.overview.delete')}
-            </Button>
-          </Inline>
-        </Inline>
-
-        <Stack spacing={5}>
-          <Stack>
-            <Text color={colors.udbMainDarkBlue} fontWeight="bold">
-              {t('create.calendar.fixed_days.overview.weekly_on')}
-            </Text>
-            <Stack>
-              {openingHours.map((openingHour, index) => (
-                <Stack
-                  key={`${openingHour.opens}-${openingHour.closes}-${openingHour.dayOfWeek.join(',')}`}
-                  paddingY={2}
-                  css={`
-                    ${index > 0 ? `border-top: 1px solid ${colors.grey3};` : ''}
-                  `}
-                >
-                  <Inline alignItems="flex-start">
-                    <Text minWidth="12rem">
-                      {openingHour.dayOfWeek
-                        .map((day) => t(`create.calendar.days.short.${day}`))
-                        .join(', ')}
-                    </Text>
-                    <Text minWidth="8rem">
-                      {openingHour.opens} - {openingHour.closes}
-                    </Text>
-                    <Stack spacing={1} minWidth="10rem">
-                      {openingHour.childcareStartTime &&
-                        openingHour.childcareEndTime && (
-                          <>
-                            <Text fontStyle="italic">
-                              {t(
-                                'create.calendar.fixed_days.overview.childcare_before',
-                                { start: openingHour.childcareStartTime },
-                              )}
-                            </Text>
-                            <Text fontStyle="italic">
-                              {t(
-                                'create.calendar.fixed_days.overview.childcare_after',
-                                { end: openingHour.childcareEndTime },
-                              )}
-                            </Text>
-                          </>
-                        )}
-                    </Stack>
-                  </Inline>
-                </Stack>
-              ))}
-            </Stack>
-          </Stack>
-
-          {initialAdjustedDays && initialAdjustedDays.length > 0 && (
-            <Stack>
-              <Text color={colors.udbMainDarkBlue} fontWeight="bold">
-                {t('create.calendar.fixed_days.overview.deviating_except')}
-              </Text>
-              <Stack spacing={4}>
-                {initialAdjustedDays.map((period) => (
-                  <Stack key={period.id} spacing={1}>
-                    <Text>
-                      {`${format(period.startDate, 'dd/MM/yyyy')} - ${format(period.endDate, 'dd/MM/yyyy')}`}
-                      {period.description[lang]
-                        ? ` (${period.description[lang]})`
-                        : ''}
-                    </Text>
-                    <Text color={colors.udbMainDarkBlue}>
-                      {t(
-                        'create.calendar.fixed_days.overview.deviating_then_weekly',
-                      )}
-                    </Text>
-                    {period.openingHours.map((openingHour, index) => (
-                      <Stack
-                        key={`${openingHour.opens}-${openingHour.closes}-${openingHour.dayOfWeek.join(',')}`}
-                        paddingTop={index > 0 ? 2 : 0}
-                        paddingBottom={2}
-                        css={`
-                          ${index > 0
-                            ? `border-top: 1px solid ${colors.grey3};`
-                            : ''}
-                        `}
-                      >
-                        <Inline alignItems="flex-start">
-                          <Text minWidth="12rem">
-                            {openingHour.dayOfWeek
-                              .map((day) =>
-                                t(`create.calendar.days.short.${day}`),
-                              )
-                              .join(', ')}
-                          </Text>
-                          <Text minWidth="8rem">
-                            {openingHour.opens} - {openingHour.closes}
-                          </Text>
-                          <Stack spacing={1} minWidth="10rem">
-                            {openingHour.childcare?.start &&
-                              openingHour.childcare?.end && (
-                                <>
-                                  <Text fontStyle="italic">
-                                    {t(
-                                      'create.calendar.fixed_days.overview.childcare_before',
-                                      { start: openingHour.childcare.start },
-                                    )}
-                                  </Text>
-                                  <Text fontStyle="italic">
-                                    {t(
-                                      'create.calendar.fixed_days.overview.childcare_after',
-                                      { end: openingHour.childcare.end },
-                                    )}
-                                  </Text>
-                                </>
-                              )}
-                          </Stack>
-                        </Inline>
-                      </Stack>
-                    ))}
-                  </Stack>
-                ))}
-              </Stack>
-            </Stack>
-          )}
-
-          {initialClosingPeriods && initialClosingPeriods.length > 0 && (
-            <Stack spacing={1}>
-              <Text color={colors.udbMainDarkBlue} fontWeight="bold">
-                {t('create.calendar.fixed_days.overview.closed')}
-              </Text>
-              {initialClosingPeriods.map((period) => (
-                <Text key={period.id}>
-                  {`${format(period.startDate, 'dd/MM/yyyy')} - ${format(period.endDate, 'dd/MM/yyyy')}`}
-                  {period.description[lang]
-                    ? ` (${period.description[lang]})`
-                    : ''}
-                </Text>
-              ))}
-            </Stack>
-          )}
-        </Stack>
-      </Stack>
-    ) : hasOpeningHours ? (
-      <List>
-        <List.Item
-          alignItems="center"
-          paddingTop={3}
-          paddingBottom={3}
-          justifyContent="space-between"
-          spacing={5}
-        >
-          <Text fontWeight="bold">
-            {t('create.calendar.fixed_days.opening_hours')}
-          </Text>
-          <Button
-            variant={ButtonVariants.SECONDARY}
-            onClick={() => setIsCalendarOpeninghoursModalVisible(true)}
-          >
-            {t('create.calendar.fixed_days.button_change_opening_hours')}
-          </Button>
-        </List.Item>
-        {openingHours.map((openingHour) => (
-          <List.Item
-            paddingTop={2}
-            paddingBottom={2}
-            css={`
-              border-top: 1px solid ${colors.grey3};
-            `}
-            justifyContent="space-between"
-            spacing={2}
-            key={`${openingHour.opens}-${openingHour.closes}-${openingHour.dayOfWeek.join(',')}`}
-          >
-            <Text maxWidth="20rem">
-              {openingHour.dayOfWeek
-                .map((dayOfWeek) =>
-                  t(`create.calendar.days.short.${dayOfWeek}`),
-                )
-                .join(', ')}
-            </Text>
-            <Text>
-              {openingHour.opens} - {openingHour.closes}
-            </Text>
-          </List.Item>
-        ))}
-      </List>
-    ) : (
-      <Button
-        variant={ButtonVariants.SECONDARY}
-        onClick={() => setIsCalendarOpeninghoursModalVisible(true)}
-        alignSelf="flex-start"
-      >
-        {t(
-          isBoaEnabled
-            ? 'create.calendar.fixed_days.button_add_hours'
-            : 'create.calendar.fixed_days.button_add_opening_hours',
-        )}
-      </Button>
-    );
 
   return (
     <Stack spacing={5} alignItems="flex-start">
@@ -362,7 +130,12 @@ export const FixedDays = ({
               apiHolidays={apiHolidays}
               onShowHolidaysChange={onShowHolidaysChange}
             />
-            {openingHoursContent}
+            <OpeningHoursContent
+              initialAdjustedDays={initialAdjustedDays}
+              initialClosingPeriods={initialClosingPeriods}
+              onOpenModal={() => setIsCalendarOpeninghoursModalVisible(true)}
+              onRequestDelete={() => setIsDeleteConfirmVisible(true)}
+            />
           </Stack>
         )}
         <RadioButtonWithLabel
@@ -373,7 +146,16 @@ export const FixedDays = ({
           onChange={handleChangeOption}
           label={t('create.calendar.fixed_days.permanent')}
         />
-        {isPermanent && <Stack paddingX={4.5}>{openingHoursContent}</Stack>}
+        {isPermanent && (
+          <Stack paddingX={4.5}>
+            <OpeningHoursContent
+              initialAdjustedDays={initialAdjustedDays}
+              initialClosingPeriods={initialClosingPeriods}
+              onOpenModal={() => setIsCalendarOpeninghoursModalVisible(true)}
+              onRequestDelete={() => setIsDeleteConfirmVisible(true)}
+            />
+          </Stack>
+        )}
       </Stack>
       {!isBoaEnabled && (
         <CalendarOpeninghoursModalLegacy
